+++ FinalTask/src/main/java/by/training/final_task/service/interfaces/CompanyProviderService.java
@@ -6,7 +6,7 @@ import by.training.final_task.service.ServiceException;
 import java.util.List;
 
 public interface CompanyProviderService extends Service {
-    CompanyProvider get(int id) throws ServiceException;
+    CompanyProvider get(long id) throws ServiceException;
     CompanyProvider getByPhone(int phone) throws ServiceException;
     List<CompanyProvider> getAllAvailableCompany(int offset, int limit)
         throws ServiceException;
